cleanup imap impl
