+++ src/lib.rs
@@ -40,6 +40,7 @@ pub mod contact;
 pub mod context;
 mod e2ee;
 mod imap;
+mod imap_client;
 pub mod imex;
 pub mod job;
 mod job_thread;
